# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Protocolo Cirúrgico Ortognático — Clínica da Face
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /protocols/{id}/planning-images:
    get:
      operationId: listPlanningImages
      tags: [planning]
      summary: List planning images for a protocol
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Planning images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanningImage"
        "404":
          description: Protocol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: createPlanningImage
      tags: [planning]
      summary: Register a planning image after upload
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanningImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningImage"
        "404":
          description: Protocol not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /protocols/{id}/planning-images/reorder:
    post:
      operationId: reorderPlanningImages
      tags: [planning]
      summary: Reorder planning images
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanningImage"
  /protocols/{id}/planning-images/{imageId}:
    patch:
      operationId: updatePlanningImage
      tags: [planning]
      summary: Update caption, category or order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanningImageUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningImage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePlanningImage
      tags: [planning]
      summary: Delete a planning image
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /plate-catalog:
    get:
      operationId: listPlateCatalog
      tags: [catalog]
      summary: List plate catalogue entries
      parameters:
        - name: manufacturer
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Catalogue entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlateCatalogEntry"
    post:
      operationId: createPlateCatalogEntry
      tags: [catalog]
      summary: Create a plate catalogue entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlateCatalogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlateCatalogEntry"
  /plate-catalog/{entryId}:
    patch:
      operationId: updatePlateCatalogEntry
      tags: [catalog]
      summary: Update a plate catalogue entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlateCatalogUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlateCatalogEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deletePlateCatalogEntry
      tags: [catalog]
      summary: Delete a plate catalogue entry
      parameters:
        - name: entryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── Protocol schemas ───────────────────────────────────────
...
    PlanningImage:
      type: object
      required: [id, protocolId, objectPath, servingUrl, category, displayOrder, createdAt, updatedAt]
      properties:
        id: { type: integer }
        protocolId: { type: integer }
        objectPath: { type: string }
        servingUrl: { type: string }
        originalName: { type: ["string", "null"] }
        caption: { type: ["string", "null"] }
        category:
          type: string
          enum: [foto_extraoral, foto_intraoral, foto_clinica_outra, fotografias_clinicas, cefalometria, renders_3d, tecidos_moles, osteotomias, movimentos_osseos, splints_guias, comparacao_pre_pos, outros]
        phase: { type: ["string", "null"] }
        captureDate: { type: ["string", "null"] }
        isHeaderPhoto: { type: boolean }
        displayOrder: { type: integer }
        includeInPdf: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    PlanningImageInput:
      type: object
      required: [objectPath, servingUrl, category]
      properties:
        objectPath: { type: string }
        servingUrl: { type: string }
        originalName: { type: string }
        caption: { type: string }
        category:
          type: string
          enum: [foto_extraoral, foto_intraoral, foto_clinica_outra, fotografias_clinicas, cefalometria, renders_3d, tecidos_moles, osteotomias, movimentos_osseos, splints_guias, comparacao_pre_pos, outros]
        phase: { type: string }
        captureDate: { type: string }
        isHeaderPhoto: { type: boolean }
        displayOrder: { type: integer }
        includeInPdf: { type: boolean }
    PlanningImageUpdate:
      type: object
      properties:
        caption: { type: string }
        category:
          type: string
          enum: [foto_extraoral, foto_intraoral, foto_clinica_outra, fotografias_clinicas, cefalometria, renders_3d, tecidos_moles, osteotomias, movimentos_osseos, splints_guias, comparacao_pre_pos, outros]
        phase: { type: string }
        captureDate: { type: string }
        isHeaderPhoto: { type: boolean }
        displayOrder: { type: integer }
        includeInPdf: { type: boolean }

    # ─── 3D file schemas ─────────────────────────────────────────
...
    PlateCatalogEntry:
      type: object
      required: [id, manufacturer, reference, active, createdAt, updatedAt]
      properties:
        id: { type: integer }
        manufacturer:
          type: string
          enum: [osteomed, delphos, outro]
        system: { type: ["string", "null"] }
        plateType: { type: ["string", "null"] }
        geometry: { type: ["string", "null"] }
        side: { type: ["string", "null"] }
        region: { type: ["string", "null"] }
        reference: { type: string }
        screwInfo: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        active: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    PlateCatalogInput:
      type: object
      required: [manufacturer, reference]
      properties:
        manufacturer:
          type: string
          enum: [osteomed, delphos, outro]
        system: { type: string }
        plateType: { type: string }
        geometry: { type: string }
        side: { type: string }
        region: { type: string }
        reference: { type: string }
        screwInfo: { type: string }
        notes: { type: string }
        active: { type: boolean }
    PlateCatalogUpdate:
      type: object
      properties:
        manufacturer:
          type: string
          enum: [osteomed, delphos, outro]
        system: { type: string }
        plateType: { type: string }
        geometry: { type: string }
        side: { type: string }
        region: { type: string }
        reference: { type: string }
        screwInfo: { type: string }
        notes: { type: string }
        active: { type: boolean }
    ReorderInput:
      type: object
      required: [orderedIds]
      properties:
        orderedIds:
          type: array
          items:
            type: integer
...